// lib/constants/enums.ts
export enum OpenDEXnetwork {
  // real coins.
  MainNet = 'mainnet',

  // pre-defined testnet per currency (e.g. bitcoin: testnet3, litecoin: testnet4).
  TestNet = 'testnet',

  // coins are minted by Exchange Union.
  SimNet = 'simnet',

  // coins are minted privately. Seed nodes are not configurable.
  RegTest = 'regtest',
}